scripts/search/keywords/requirements.txt
keybert>=0.9.0
ruamel.yaml>=0.18.10
frontmatter>=3.0.8